Use encodeURLPart in AuthorizationHeaderBuilder, fix some encoding issues
